requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.12
arrow==1.2.3
astroid==2.14.2
asttokens==2.0.5
async-timeout==4.0.3
asyncssh==2.17.0
atomicwrites==1.4.0
attrs==23.1.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
BERTSimilarity==0.1
binaryornot==0.4.4
black==0.0
bleach==4.1.0
blinker==1.8.2
Brotli==1.0.9
certifi==2024.7.4
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
cookiecutter==1.7.3
cryptography==41.0.3
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.15
diff-match-patch==20200713
dill==0.3.7
docstring-to-markdown==0.11
docutils==0.18.1
exceptiongroup==1.0.4
executing==0.8.3
fairscale==0.4.13
fastjsonschema==2.16.2
filelock==3.9.0
flake8==6.0.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==2.10
imagesize==1.4.1
importlib_metadata==8.5.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.25.0
ipython==8.15.0
isort==5.9.3
itsdangerous==2.2.0
jaraco.classes==3.2.1
jedi==0.18.1
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyterlab-pygments==0.1.2
keyring==23.13.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
more-itertools==8.12.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numpy==1.24.3
numpydoc==1.5.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
pip==23.3
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.36
propcache==0.2.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
PyGithub==2.4.0
Pygments==2.15.1
PyJWT==2.10.1
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
PyNaCl==1.5.0
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtWebEngine==5.15.4
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-slugify==5.0.2
pytoolconfig==1.2.5
pytz==2023.3.post1
pyuca==1.2
pywin32==305.1
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.2
QtPy==2.2.0
referencing==0.30.2
regex==2023.10.3
requests==2.32.3
rich==13.9.4
rope==1.7.0
rpds-py==0.10.6
Rtree==1.0.1
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
setuptools==68.0.0
sip==6.6.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.3
spyder-kernels==2.4.4
stack-data==0.2.0
superqt==0.6.7
sympy==1.12
tabulate==0.9.0
text-unidecode==1.3
textblob==0.18.0.post0
textdistance==4.2.1
threadpoolctl==3.2.0
three-merge==0.1.1
timm==1.0.8
tinycss2==1.2.1
tokenizers==0.14.1
tomli==2.0.1
tomlkit==0.11.1
torch==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.66.5
traitlets==5.12.0
transformers==4.35.0
typing_extensions==4.12.2
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.18
watchdog==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
whatthepatch==1.0.2
wheel==0.41.2
win-inet-pton==1.1.0
wrapt==1.14.1
yapf==0.31.0
yarl==1.9.4
zipp==3.11.0
